refactor(oauth): split oauth-manager.ts into types, core, and barrel
Extract types/interfaces into oauth-types.ts (OAuthPlatform, OAuthConfig,
OAuthTokenResponse, AuthorizationUrlResponse, OAuthStatus).
Move OAuthManager class and singleton helpers into oauth-manager-core.ts.
Keep oauth-manager.ts as a barrel re-export for backward compatibility.

Refs: #1

Cost-free: ✅
